requirements/common.txt
fastapi==0.105.0
uvicorn==0.20.0
sqlalchemy==2.0.3
pydantic==2.8.2
python-dotenv==1.0.1
alembic==1.9.2
psycopg2-binary==2.9.5
gunicorn==20.1.0
pynamodb==6.0.0
boto3==1.26.118
celery[sqs]>=5.2.7, <6
tenacity>=8.0.0, <9
requests==2.30.0
PyJWT==2.10.1